+++ float-preload.cjs
@@ -1,0 +1,5 @@
+const { contextBridge, ipcRenderer } = require('electron');
+
+contextBridge.exposeInMainWorld('sidebar', {
+  quit: () => ipcRenderer.send('sidebar-quit'),
+});